requirements.txt
# Core Framework
fastapi==0.115.14
uvicorn>=0.15.0
pydantic[email]==2.11.7
pydantic-settings==2.10.1
starlette>=0.27.0

# Database - PostgreSQL (Neon)
sqlmodel>=0.0.8
psycopg2-binary>=2.9.1
asyncpg==0.29.0
alembic==1.13.1
greenlet>=3.0.0

# Security & Authentication
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
authlib==1.3.0
bcrypt==4.1.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
PyJWT==2.8.0

# Payments
stripe==7.12.0

# HTTP & Async
httpx==0.26.0
requests>=2.26.0
python-multipart>=0.0.5,<0.0.6

# External Services
openai>=1.0.0
twilio==8.12.0

# Utilities
python-dotenv>=0.19.0
python-slugify==8.0.1
jinja2==3.1.2
tenacity==8.2.3
itsdangerous==2.1.2

# Scheduling
APScheduler==3.10.4

# Security & Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Monitoring & Logging
structlog>=23.1.0
prometheus-client==0.19.0
python-json-logger>=2.0.7

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.2
black==23.12.1
ruff==0.1.8
